+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 (Add entries below in the order of server, console, cli, docs, others)
 
+- server: fix an issue with remote relationships when join columns are aliased (close #7180)
 - server: fix for incorrect `__typename` value in nested remote joins with a customized remote schema
 - server: fix a bug where some unicode characters in default string values for fields in remote schemas could lead to internal errors
 - console: fix untracked foreign-key relationships suggestion across schemas